refactor: reorganize Supabase client and database types for improved structure
