+++ lib/supabase/supabase.client.ts
@@ -1,6 +1,6 @@
 import { auth } from "@clerk/nextjs/server";
 import { createClient } from "@supabase/supabase-js";
-import { Database } from "./database.types";
+import { Database } from "../../types/database";
 
 export function createSupabaseClient() {
   return createClient<Database, "lms_saas_app">(
